Transfer `ExtendedDataFormat` to `data.rs`
`ExtendedDataFormat` is used only by `Data128` which is defined in `data.rs`
