+++ src/message/data.rs
@@ -1,5 +1,6 @@
 use core::ops::Deref;
 
+use crate::message::DataFormat;
 use crate::packet::{MessageType, Packet128, Packet64};
 
 use super::Message;
@@ -9,16 +10,6 @@ pub struct Data64(Packet64);
 
 #[derive(Copy, Clone, Hash, Debug, Eq, PartialEq)]
 pub struct Data128(Packet128);
-
-#[derive(Copy, Clone, Hash, Debug, Eq, PartialEq)]
-pub enum DataStatus {
-    SinglePacket = 0x0,
-    Start = 0x1,
-    Continue = 0x2,
-    End = 0x3,
-    MixedDataSetHeader = 0x8,
-    MixedDataSetPayload = 0x9,
-}
 
 impl Data64 {
     pub(crate) fn from_packet_unchecked(ump: Packet64) -> Self {
@@ -34,7 +25,7 @@ impl Deref for Data64 {
 }
 
 impl Message for Data64 {
-    type Status = DataStatus;
+    type Status = DataFormat;
     type Data = [u8; 6];
 
     fn message_type(&self) -> MessageType {
@@ -48,24 +39,18 @@ impl Message for Data64 {
     }
 
     fn status(&self) -> Self::Status {
-        match (self.0[0] >> 20) & 0xf {
-            0x0 => Self::Status::SinglePacket,
-            0x1 => Self::Status::Start,
-            0x2 => Self::Status::Continue,
-            0x3 => Self::Status::End,
-            _ => unreachable!("Invalid status byte for 8 byte data message."),
-        }
+        (((self.0[0] >> 20) & 0xf) as u8).into()
     }
 
     fn data(&self) -> Self::Data {
         let bytes = [self.0[0].to_ne_bytes(), self.0[1].to_ne_bytes()];
         [
-            bytes[0][1],
             bytes[0][0],
-            bytes[1][3],
-            bytes[1][2],
-            bytes[1][1],
+            bytes[0][1],
             bytes[1][0],
+            bytes[1][1],
+            bytes[1][2],
+            bytes[1][3],
         ]
     }
 }
@@ -96,7 +81,7 @@ impl Data128 {
 }
 
 impl Message for Data128 {
-    type Status = DataStatus;
+    type Status = ExtendedDataFormat;
     type Data = [u8; 14];
 
     fn message_type(&self) -> MessageType {
@@ -110,15 +95,7 @@ impl Message for Data128 {
     }
 
     fn status(&self) -> Self::Status {
-        match (self.0[0] >> 20) & 0xf {
-            0x0 => Self::Status::SinglePacket,
-            0x1 => Self::Status::Start,
-            0x2 => Self::Status::Continue,
-            0x3 => Self::Status::End,
-            0x4 => Self::Status::MixedDataSetHeader,
-            0x5 => Self::Status::MixedDataSetPayload,
-            _ => unreachable!("Invalid status byte for 16 byte data message."),
-        }
+        (((self.0[0] >> 20) & 0xf) as u8).into()
     }
 
     fn data(&self) -> Self::Data {
@@ -129,21 +106,61 @@ impl Message for Data128 {
             self.0[3].to_ne_bytes()
         ];
         [
-            bytes[0][1],
             bytes[0][0],
-            bytes[1][3],
-            bytes[1][2],
-            bytes[1][1],
+            bytes[0][1],
             bytes[1][0],
-            bytes[2][3],
-            bytes[2][2],
-            bytes[2][1],
+            bytes[1][1],
+            bytes[1][2],
+            bytes[1][3],
             bytes[2][0],
-            bytes[3][3],
-            bytes[3][2],
-            bytes[3][1],
+            bytes[2][1],
+            bytes[2][2],
+            bytes[2][3],
             bytes[3][0],
+            bytes[3][1],
+            bytes[3][2],
+            bytes[3][3],
         ]
+    }
+}
+
+
+#[derive(Copy, Clone, Hash, Debug, Eq, PartialEq)]
+pub enum ExtendedDataFormat {
+    SinglePacket,
+    Start,
+    Continue,
+    End,
+    MixedDataSetHeader,
+    MixedDataSetPayload,
+    Reserved,
+}
+
+impl From<u8> for ExtendedDataFormat {
+    fn from(value: u8) -> Self {
+        match value {
+            0 => Self::SinglePacket,
+            1 => Self::Start,
+            2 => Self::Continue,
+            3 => Self::End,
+            8 => Self::MixedDataSetHeader,
+            9 => Self::MixedDataSetPayload,
+            _ => Self::Reserved,
+        }
+    }
+}
+
+impl From<ExtendedDataFormat> for u8 {
+    fn from(value: ExtendedDataFormat) -> Self {
+        match value {
+            ExtendedDataFormat::SinglePacket => 0,
+            ExtendedDataFormat::Start => 1,
+            ExtendedDataFormat::Continue => 2,
+            ExtendedDataFormat::End => 3,
+            ExtendedDataFormat::MixedDataSetHeader => 8,
+            ExtendedDataFormat::MixedDataSetPayload => 9,
+            ExtendedDataFormat::Reserved => unreachable!()
+        }
     }
 }
